% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmos_table.R
\name{atmos_table}
\alias{atmos_table}
\title{Tabulates into a tibble some properties of the standard atmosphere: temperature, 
density, and pressure.}
\usage{
atmos_table(units = c("SI", "Eng"), ret_units = TRUE)
}
\arguments{
\item{units}{character vector that contains the system of units [options are
\code{SI} for International System of Units and \code{Eng} for English (US customary)
units.}

\item{ret_units}{If set to TRUE the value(s) returned are of class \code{units} with
units attached to the value. [Default is TRUE]}
}
\description{
Tabulates into a tibble some properties of the standard atmosphere: temperature, 
density, and pressure.
}
\examples{

atmos_table(units = 'SI')

}
\author{
Ed Maurer
}
