% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration.r
\name{calc_dTheta}
\alias{calc_dTheta}
\title{change in the moisture content when the wetting front passes}
\usage{
calc_dTheta(se, theta_r, porosity)
}
\arguments{
\item{se}{is the effective saturation and can be obtained by function set}

\item{theta_r}{is the residual moisture content of the soil after it has been thoroughly drained (in this case a soil parameter that can be obtained from a table)}

\item{porosity}{is the porosity of the soil}
}
\description{
change in the moisture content when the wetting front passes
}
