% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration.r
\name{case2}
\alias{case2}
\title{cumulative infiltration rate *F*
if ft is greater than i_t and no ponding occurs during the interval}
\usage{
case2(Ft0, dt, i_t)
}
\description{
cumulative infiltration rate *F*
if ft is greater than i_t and no ponding occurs during the interval
}
