% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration.r
\name{case3}
\alias{case3}
\title{cumulative infiltration rate *F*
if ft is less than or equal to i_t and ponding occurs during the interval}
\usage{
case3(Ft0, ft0, phi, dTheta, K, dt, i_t)
}
\arguments{
\item{phi}{wetting front capillary pressure head (parameter see chow page 115)}

\item{dTheta}{difference porosity and initial soil moisture content}

\item{K}{hydraulic conductivity of the soil (parameter see chow page 115)}

\item{F0}{cumulative infiltration. Obtain from initial conditions or previous iteration}
}
\description{
cumulative infiltration rate *F*
if ft is less than or equal to i_t and ponding occurs during the interval
}
