% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsurface_flow.r
\name{F}
\alias{F}
\title{cumulative infiltration fate *F*}
\usage{
F(K, t, phi, dTheta, F_)
}
\arguments{
\item{K}{hydraulic conductivity of the soil (parameter see chow page 115)}

\item{phi}{wetting front capillary pressure head (parameter see chow page 115)}

\item{dTheta}{difference between initial and final moisture contents of the soil}

\item{F_}{cumulative infiltration trial value (it must converge until F==F_)}
}
\description{
cumulative infiltration fate *F*
}
