% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_runoff.r
\name{inlet_time}
\alias{inlet_time}
\title{time of concentration upstream of sewers in minutes, based on the kinematic wave, chow 501}
\usage{
inlet_time(L, n, i, S)
}
\arguments{
\item{L}{length of overland flow, m}

\item{n}{is the manning roughness coefficient}

\item{i}{is the rainfall intensity, mm/h}

\item{S}{average overland slope m/m}
}
\description{
time of concentration upstream of sewers in minutes, based on the kinematic wave, chow 501
}
