% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operpoint.R
\name{operpoint}
\alias{operpoint}
\title{Uses input pump and system curves to find the operating point
for a pump and create a plot.}
\usage{
operpoint(pcurve = NULL, scurve = NULL)
}
\arguments{
\item{pcurve}{A pumpcurve object}

\item{scurve}{A systemcurve object}
}
\value{
Returns a list including:
\itemize{
  \item Qop - flow at the operating point [\eqn{m^3 s^{-1}}{m^3/s} or \eqn{ft^3 s^{-1}}{ft^3/s}]
  \item hop - head at the operating point [\eqn{m}{m} or \eqn{ft}{ft}]
  \item p - a plot object of the curves
}
}
\description{
Uses input pump and system curves to find the operating point
for a pump and create a plot.
}
\seealso{
\code{\link{pumpcurve}} and \code{\link{systemcurve}} for input object preparation
}
\author{
Ed Maurer
}
