% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsurface_flow.r
\name{phi}
\alias{phi}
\title{brooks-correy equation for wetting front capillary pressure head}
\usage{
phi(se, phi_b, lambda)
}
\arguments{
\item{se}{is the effective saturation given above}

\item{phi_b}{is obtained in lab by draining a soil in stages}

\item{lambda}{is obtained in lab by draining a soil in stages}
}
\description{
brooks-correy equation for wetting front capillary pressure head
}
