% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_routing.r
\name{Q_out_muskingum}
\alias{Q_out_muskingum}
\title{Muskingum's effluent in pipe or catchment reach as in the documentation of citydrain 2}
\usage{
Q_out_muskingum(state)
}
\arguments{
\item{state}{is a data frame of colums: Qin column of inflows to reach in step i; Vprevious is the column of volumes in reach or basin in step i-1; Qout is the column of outflows from reach in step i; dt, K, X are timestep, K and X as in muskingum}
}
\description{
Muskingum's effluent in pipe or catchment reach as in the documentation of citydrain 2
}
