% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention_structure_df.r
\name{Virtual_retention}
\alias{Virtual_retention}
\title{Virtual volume in retention structure}
\usage{
Virtual_retention(ret_str_state)
}
\arguments{
\item{ret_str_state}{is a dataframe with columns:}

\item{Qin}{is a column with inflow to structure in step i}

\item{Qoverflow}{is a column unregulated overflow from structure in step i}

\item{Vvirtual}{is the virtual volume in step i, including overflow volume}

\item{Vprevious}{is the volume in step i-1}

\item{dt}{is a constant column (time step)}

\item{Vmax}{is a constant column (maximum allowable volume of structure)}

\item{Qoutmax}{is a constant column (maximum regulated allowable outflow from structure)}
}
\description{
Virtual volume in retention structure
}
