% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_table.R
\name{water_table}
\alias{water_table}
\title{Tabulates into a tibble the basic water properties: density, dynamic and kinematic 
viscosity, saturation vapor pressure, surface tension, and bulk modulus.}
\usage{
water_table(units = c("SI", "Eng"), ret_units = TRUE)
}
\arguments{
\item{units}{character vector that contains the system of units [options are
\code{SI} for International System of Units and \code{Eng} for English (US customary)
units. This is used for compatibility with iemisc package}

\item{ret_units}{If set to TRUE the value(s) returned are of class \code{units} with
units attached to the value. [Default is TRUE]}
}
\description{
Tabulates into a tibble the basic water properties: density, dynamic and kinematic 
viscosity, saturation vapor pressure, surface tension, and bulk modulus.
}
\examples{

water_table(units = 'SI')

}
\author{
Ed Maurer
}
