% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_circle.R
\name{xc_circle}
\alias{xc_circle}
\title{Creates a cross-section plot for a partially filled pipe}
\usage{
xc_circle(y = NULL, d = NULL)
}
\arguments{
\item{y}{water depth}

\item{d}{pipe diameter}
}
\value{
a cross-section diagram
}
\description{
This function plots a cross-section of a circular pipe, shaded as filled
to the level indicated by the depth and diameter values passed to it.
}
\examples{

# Draw a cross-section with diameter 1.0 and depth 0.7
xc_circle(y = 0.7, d = 1.0)

}
\author{
Ed Maurer
}
