% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_trap.R
\name{xc_trap}
\alias{xc_trap}
\title{Creates a cross-section plot for an open channel}
\usage{
xc_trap(y = NULL, b = NULL, m = NULL)
}
\arguments{
\item{y}{water depth}

\item{b}{bottom width}

\item{m}{side slope (H:1)}
}
\value{
a cross-section diagram
}
\description{
This function plots a cross-section of a  (trapezoid, rectangle, triangle), 
shaded as filled to the level indicated by the values passed to it.
}
\examples{

# Draw a cross-section with depth 1, width 2, side slope 3:1 (H:V)
xc_trap(y = 1.0, b = 2.0, m = 3.0)

}
\author{
Ed Maurer
}
